asset basics
